fix(ch25): route trace detail through the v1/v2 dispatch (V1=PG, V2=CH)
GET /tracer/trace/{id}/ 500s on collector-ingested (CH-only) traces because
retrieve() did a hard PG Trace lookup that has no row for them. Route the
endpoint through the same routing-mode dispatch the list queries use:

- Register TRACE_DETAIL in dispatch._REGISTRY (v1 PG handler, v2 CH handler).
- TraceDetailHandler (V1/PG): original behaviour, unchanged — kept for
  stability so existing deployments are unaffected until flipped to V2.
- TraceDetailHandlerV2 (V2/CH): resolves project from CH spans, CH-based
  tenant gate, synthesizes trace metadata from the root span when no PG row.
- trace.py::retrieve() now delegates to the dispatched handler's fetch();
  the ~390-line inline PG/CH bodies move into the handler classes.

[TH-5964]
