+++ futureagi/tracer/services/clickhouse/v2/dispatch.py
@@ -102,6 +102,12 @@ _REGISTRY: dict[str, _BuilderEntry] = {
         v2_module="tracer.services.clickhouse.v2.query_builders.filters",
         v2_class="ClickHouseFilterBuilderV2",
     ),
+    "TRACE_DETAIL": _BuilderEntry(
+        v1_module="tracer.services.clickhouse.query_builders.trace_detail",
+        v1_class="TraceDetailHandler",
+        v2_module="tracer.services.clickhouse.v2.query_builders.trace_detail",
+        v2_class="TraceDetailHandlerV2",
+    ),
 }
