+++ futureagi/tracer/views/trace.py
@@ -68,6 +68,7 @@ from tracer.serializers.trace import (
     UsersQuerySerializer,
     UsersResponseSerializer,
 )
+from tracer.services.clickhouse.eval_logger_table import eval_logger_source
 from tracer.services.clickhouse.graph_dispatch import (
     fetch_annotation_graph_ch,
     fetch_eval_graph_ch,
@@ -77,7 +78,6 @@ from tracer.services.clickhouse.query_builders import (
     AgentGraphQueryBuilder,
     UserListQueryBuilder,
 )
-from tracer.services.clickhouse.eval_logger_table import eval_logger_source
 from tracer.services.clickhouse.query_builders.base import NIL_UUID
 from tracer.services.clickhouse.query_service import AnalyticsQueryService
 from tracer.services.observability_providers import ObservabilityService
@@ -1109,20 +1109,18 @@ class TraceView(BaseModelViewSetMixin, ModelViewSet):
         """
         try:
             trace_id = kwargs.get("pk")
-            # Cross-store tenant gate (get_queryset filters by org/workspace).
-            accessible_trace = self.get_queryset().filter(id=trace_id).first()
-            if not accessible_trace:
-                raise Trace.DoesNotExist
+            from tracer.services.clickhouse.v2.dispatch import (
+                get_query_builder_class,
+            )
 
-            # CH-only path post-migration. D-027: the previous PG fallback
-            # (get_observation_spans + _compute_summary_and_graph against
-            # the PG tracer_observation_span table) was deleted. CH is the
-            # authoritative span store; _retrieve_clickhouse produces the
-            # span tree, summary, and graph from the CH `spans` table. If
-            # TRACE_DETAIL isn't routed to CH that's a config error —
-            # surface it as a 400 rather than serving partial PG data.
-            analytics = AnalyticsQueryService()
-            return self._retrieve_clickhouse(request, trace_id, analytics)
+            HandlerCls = get_query_builder_class("TRACE_DETAIL")
+            handler = HandlerCls(
+                view=self,
+                request=request,
+                pk=trace_id,
+                analytics=AnalyticsQueryService(),
+            )
+            return self._gm.success_response(handler.fetch())
         except Exception as e:
             logger.exception(f"Error in fetching the trace: {str(e)}")
             return self._gm.bad_request(
@@ -1201,387 +1199,6 @@ class TraceView(BaseModelViewSetMixin, ModelViewSet):
         }
         graph = {"nodes": graph_nodes, "edges": graph_edges}
         return summary, graph
-
-    def _retrieve_clickhouse(self, request, trace_id, analytics):
-        """Retrieve a trace and its spans from ClickHouse."""
-        from tracer.constants.provider_logos import PROVIDER_LOGOS
-
-        # Always fetch trace metadata from PG (small, config-like)
-        trace = Trace.objects.get(
-            id=trace_id,
-            project__organization=getattr(self.request, "organization", None)
-            or self.request.user.organization,
-        )
-        serializer = self.get_serializer(trace)
-        trace_data = serializer.data
-
-        project_id = str(trace.project_id)
-
-        # Fetch all spans for this trace from CH — use the denormalized `spans`
-        # table which has renamed columns vs PG. Map them back to expected names.
-        query = """
-            SELECT
-                id, trace_id, parent_span_id, name, observation_type,
-                start_time, end_time, input, output, model,
-                '' AS model_parameters, latency_ms, prompt_tokens,
-                completion_tokens, total_tokens, cost, status,
-                status_message, tags, span_events,
-                provider, attributes_extra AS span_attributes,
-                project_version_id, custom_eval_config_id,
-                toJSONString(metadata) AS metadata_json,
-                attrs_string, attrs_number, attrs_bool
-            FROM spans
-            WHERE project_id = %(project_id)s
-              AND trace_id = %(trace_id)s
-              AND is_deleted = 0
-            ORDER BY start_time
-            LIMIT 1 BY id
-        """
-        result = analytics.execute_ch_query(
-            query,
-            {"project_id": project_id, "trace_id": str(trace_id)},
-            timeout_ms=10000,
-        )
-
-        # Build span tree
-        span_map = {}  # id -> span data
-        root_spans = []
-        orphan_spans = []
-
-        import json as _json
-
-        def _parse_json(val, default=None):
-            if default is None:
-                default = {}
-            if not val or not isinstance(val, str):
-                return val if val is not None else default
-            try:
-                return _json.loads(val)
-            except (ValueError, TypeError):
-                return default
-
-        for row in result.data:
-            span_id = str(row.get("id", ""))
-            parent_id = row.get("parent_span_id")
-            parent_id_str = str(parent_id) if parent_id else None
-
-            provider = row.get("provider")
-
-            # Build span_attributes from raw JSON or decomposed maps
-            span_attrs_raw = row.get("span_attributes") or "{}"
-            try:
-                span_attrs = (
-                    _json.loads(span_attrs_raw)
-                    if isinstance(span_attrs_raw, str)
-                    else span_attrs_raw
-                )
-            except (ValueError, TypeError):
-                span_attrs = {}
-            if not span_attrs:
-                span_attrs = {}
-                for k, v in (row.get("attrs_string") or {}).items():
-                    span_attrs[k] = v
-                for k, v in (row.get("attrs_number") or {}).items():
-                    span_attrs[k] = v
-                for k, v in (row.get("attrs_bool") or {}).items():
-                    span_attrs[k] = bool(v)
-            # Fallback: if CH has no span_attributes, try PG
-            if not span_attrs:
-                try:
-                    pg_span = ObservationSpan.objects.only(
-                        "span_attributes", "eval_attributes"
-                    ).get(id=span_id)
-                    span_attrs = (
-                        pg_span.span_attributes or pg_span.eval_attributes or {}
-                    )
-                except ObservationSpan.DoesNotExist:
-                    pass
-
-            # Build metadata from CH JSON column
-            metadata_raw = row.get("metadata_json") or "{}"
-            metadata = _parse_json(metadata_raw, default={})
-
-            span_data = {
-                "id": span_id,
-                "project": project_id,
-                "project_version": (
-                    str(row["project_version_id"])
-                    if row.get("project_version_id")
-                    else None
-                ),
-                "trace": str(row.get("trace_id", "")),
-                "parent_span_id": parent_id_str,
-                "name": row.get("name"),
-                "observation_type": row.get("observation_type"),
-                "start_time": row.get("start_time"),
-                "end_time": row.get("end_time"),
-                "input": _parse_json(row.get("input")),
-                "output": _parse_json(row.get("output")),
-                "model": row.get("model"),
-                "model_parameters": _parse_json(row.get("model_parameters")),
-                "latency_ms": row.get("latency_ms"),
-                "org_id": None,
-                "org_user_id": None,
-                "prompt_tokens": row.get("prompt_tokens"),
-                "completion_tokens": row.get("completion_tokens"),
-                "total_tokens": row.get("total_tokens"),
-                "response_time": None,
-                "eval_id": None,
-                "cost": (
-                    round(row["cost"], 6)
-                    if row.get("cost") and row["cost"] > 0
-                    else row.get("cost")
-                ),
-                "status": row.get("status"),
-                "status_message": row.get("status_message"),
-                "tags": _parse_json(row.get("tags"), default=[]),
-                "metadata": metadata,
-                "span_events": _parse_json(row.get("span_events"), default=[]),
-                "provider": provider,
-                "provider_logo": (
-                    PROVIDER_LOGOS.get(provider.lower()) if provider else None
-                ),
-                "span_attributes": span_attrs,
-                "custom_eval_config": (
-                    str(row["custom_eval_config_id"])
-                    if row.get("custom_eval_config_id")
-                    else None
-                ),
-                "eval_status": None,
-                "prompt_version": None,
-            }
-
-            span_map[span_id] = {
-                "observation_span": span_data,
-                "children": [],
-                "_parent_id": parent_id_str,
-            }
-
-        # ----- Phase 8: Batch fetch eval scores from CH -----
-        eval_map = {}
-        try:
-            eval_table, eval_nd = eval_logger_source()
-            eval_query = f"""
-            SELECT
-                toString(observation_span_id) AS span_id,
-                toString(custom_eval_config_id) AS eval_config_id,
-                output_float,
-                output_bool,
-                output_str,
-                eval_explanation
-            FROM {eval_table} FINAL
-            WHERE trace_id = %(trace_id)s
-              AND {eval_nd}
-            """
-            eval_result = analytics.execute_ch_query(
-                eval_query, {"trace_id": str(trace_id)}, timeout_ms=30000
-            )
-            # Collect unique config IDs for name lookup
-            config_ids_set = set()
-            for row in eval_result.data:
-                cid = row.get("eval_config_id", "")
-                if cid:
-                    config_ids_set.add(cid)
-            # Lookup eval config names from PG
-            config_lookup = {}
-            if config_ids_set:
-                configs = CustomEvalConfig.objects.filter(
-                    id__in=list(config_ids_set), deleted=False
-                ).select_related("eval_template")
-                config_lookup = {
-                    str(c.id): {
-                        # Prefer the CustomEvalConfig's user-given name (e.g.
-                        # "voice_sentence_count"), fall back to the template
-                        # name only if unset. This keeps the drawer labels in
-                        # sync with the trace list column headers.
-                        "name": c.name
-                        or (c.eval_template.name if c.eval_template else str(c.id)),
-                        "output_type": (
-                            getattr(c.eval_template, "output_type_normalized", None)
-                            if c.eval_template
-                            else None
-                        ),
-                    }
-                    for c in configs
-                }
-            # Pivot into per-span map
-            for row in eval_result.data:
-                sid = row.get("span_id", "")
-                if not sid:
-                    continue
-                if sid not in eval_map:
-                    eval_map[sid] = []
-                cid = row.get("eval_config_id", "")
-                info = config_lookup.get(cid, {})
-                # Compute score from output columns
-                output_float = row.get("output_float")
-                output_bool = row.get("output_bool")
-                output_str = row.get("output_str")
-                # Score: use float if non-zero, else bool (True=100, False=0)
-                if output_float and output_float != 0:
-                    score = round(output_float * 100, 2)
-                elif output_bool is not None:
-                    score = 100 if output_bool else 0
-                else:
-                    score = None
-
-                explanation = row.get("eval_explanation", "")
-
-                eval_map[sid].append(
-                    {
-                        "eval_config_id": cid,
-                        "eval_name": info.get("name", cid),
-                        "output_type": info.get("output_type"),
-                        "score": score,
-                        "result": output_str
-                        or (output_bool if output_bool is not None else None),
-                        "explanation": explanation if explanation else None,
-                    }
-                )
-        except Exception as e:
-            logger.warning(f"Failed to fetch trace eval scores: {e}")
-
-        # ----- Phase 8: Batch fetch annotations from PG -----
-        annotation_map = {}
-        try:
-            from model_hub.models.score import Score as ScoreModel
-
-            scores = (
-                ScoreModel.objects.filter(trace_id=trace_id, deleted=False)
-                .select_related("label")
-                .values(
-                    "observation_span_id",
-                    "label_id",
-                    "label__name",
-                    "label__type",
-                    "value",
-                )
-            )
-            for s in scores:
-                sid = (
-                    str(s["observation_span_id"])
-                    if s.get("observation_span_id")
-                    else None
-                )
-                if not sid:
-                    continue
-                if sid not in annotation_map:
-                    annotation_map[sid] = []
-                annotation_map[sid].append(
-                    {
-                        "label_id": str(s["label_id"]) if s.get("label_id") else None,
-                        "label_name": s.get("label__name"),
-                        "label_type": s.get("label__type"),
-                        "value": s.get("value"),
-                    }
-                )
-        except Exception as e:
-            logger.warning(f"Failed to fetch trace annotations: {e}")
-
-        # ----- Phase 8: Compute summary -----
-        total_tokens = 0
-        total_prompt_tokens = 0
-        total_completion_tokens = 0
-        total_cost = 0.0
-        error_count = 0
-        type_counts = {}
-        root_latencies = []
-
-        for entry in span_map.values():
-            sp = entry["observation_span"]
-            total_tokens += sp.get("total_tokens") or 0
-            total_prompt_tokens += sp.get("prompt_tokens") or 0
-            total_completion_tokens += sp.get("completion_tokens") or 0
-            total_cost += sp.get("cost") or 0.0
-            if sp.get("status") == "ERROR":
-                error_count += 1
-            obs_type = sp.get("observation_type", "unknown")
-            type_counts[obs_type] = type_counts.get(obs_type, 0) + 1
-            if entry.get("_parent_id") is None:
-                root_latencies.append(sp.get("latency_ms") or 0)
-
-        summary = {
-            "total_spans": len(span_map),
-            "total_duration_ms": max(root_latencies) if root_latencies else 0,
-            "total_tokens": total_tokens,
-            "total_prompt_tokens": total_prompt_tokens,
-            "total_completion_tokens": total_completion_tokens,
-            "total_cost": round(total_cost, 6),
-            "error_count": error_count,
-            "span_type_counts": type_counts,
-        }
-
-        # ----- Phase 8: Derive agent graph -----
-        graph_nodes = []
-        graph_edges = []
-        for sid, entry in span_map.items():
-            sp = entry["observation_span"]
-            graph_nodes.append(
-                {
-                    "id": sid,
-                    "name": sp.get("name", ""),
-                    "type": sp.get("observation_type", "unknown"),
-                    "latency_ms": sp.get("latency_ms", 0),
-                    "tokens": sp.get("total_tokens", 0),
-                    "status": sp.get("status"),
-                }
-            )
-            parent_id = entry.get("_parent_id")
-            if parent_id and parent_id in span_map:
-                graph_edges.append({"from": parent_id, "to": sid})
-
-        graph = {"nodes": graph_nodes, "edges": graph_edges}
-
-        # ----- Fetch fresh span tags from PG (CH has sync delay) -----
-        if span_map:
-            try:
-                pg_tags = dict(
-                    ObservationSpan.objects.filter(id__in=list(span_map.keys()))
-                    .exclude(tags=[])
-                    .values_list("id", "tags")
-                )
-                for sid, tags in pg_tags.items():
-                    if sid in span_map:
-                        span_map[sid]["observation_span"]["tags"] = tags
-            except Exception as e:
-                logger.warning(f"Failed to fetch span tags from PG: {e}")
-
-        # ----- Attach evals + annotations to each span -----
-        for sid, entry in span_map.items():
-            entry["eval_scores"] = eval_map.get(sid, [])
-            entry["annotations"] = annotation_map.get(sid, [])
-
-        # Build tree: link children to parents
-        for entry in span_map.values():
-            parent_id = entry["_parent_id"]
-            if parent_id is None:
-                root_spans.append(entry)
-            elif parent_id in span_map:
-                span_map[parent_id]["children"].append(entry)
-            else:
-                orphan_spans.append(entry)
-
-        # Clean up internal fields
-        def _clean_entry(entry):
-            del entry["_parent_id"]
-            for child in entry["children"]:
-                _clean_entry(child)
-
-        for entry in root_spans:
-            _clean_entry(entry)
-        for entry in orphan_spans:
-            _clean_entry(entry)
-
-        observation_spans_response = root_spans + orphan_spans
-
-        return self._gm.success_response(
-            {
-                "trace": trace_data,
-                "observation_spans": observation_spans_response,
-                "summary": summary,
-                "graph": graph,
-            }
-        )
 
     # Keys to strip from the list response (heavy / detail-only fields).
     _VOICE_CALL_HEAVY_KEYS = frozenset(
@@ -2451,9 +2068,7 @@ class TraceView(BaseModelViewSetMixin, ModelViewSet):
                             ):
                                 total_eval_configs[
                                     str(metric["custom_eval_config_id"]) + "**" + choice
-                                ] = (
-                                    metric["custom_eval_config__name"] + " - " + choice
-                                )
+                                ] = metric["custom_eval_config__name"] + " - " + choice
                 else:
                     score = (
                         metric["avg_float_score"]
